backend/requirements.txt
# Core framework
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Authentication and security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Database
supabase==1.0.4
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# API clients
httpx>=0.24.0,<0.25.0
aiohttp==3.9.1

# AI/ML
openai==1.3.8

# Image processing
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3
jmespath==1.0.1 